re-orged util
