+++ mealie/db/models/_model_utils/helpers.py
@@ -1,6 +1,5 @@
 import inspect
 from collections.abc import Callable
-from datetime import datetime, timezone
 from typing import Any
 
 
@@ -46,17 +45,3 @@ def safe_call(func, dict_args: dict | None, **kwargs) -> Any:
         return func(**get_valid_call(func, dict_args))
     except TypeError:
         return func(**dict_args)
-
-
-def get_utc_now():
-    """
-    Returns the current time in UTC.
-    """
-    return datetime.now(timezone.utc)
-
-
-def get_utc_today():
-    """
-    Returns the current date in UTC.
-    """
-    return datetime.now(timezone.utc).date()
