+++ mealie/schema/reports/reports.py
@@ -6,7 +6,7 @@ from pydantic.types import UUID4
 from sqlalchemy.orm import joinedload
 from sqlalchemy.orm.interfaces import LoaderOption
 
-from mealie.db.models._model_utils.helpers import get_utc_now
+from mealie.db.models._model_utils.datetime import get_utc_now
 from mealie.db.models.group import ReportModel
 from mealie.schema._mealie import MealieModel
